+++ backend/rpc.py
@@ -3,6 +3,7 @@ import logging
 import datetime
 
 import json
+import os
 from urllib.parse import urljoin
 from django.conf import settings
 from django.contrib.auth import authenticate, get_user_model, login as djlogin, logout as djlogout
@@ -25,8 +26,8 @@ from gatenlp import annotation_set
 from backend.errors import AuthError
 from backend.rpcserver import rpc_method, rpc_method_auth, rpc_method_manager, rpc_method_admin
 from backend.models import Project, Document, DocumentType, Annotation, AnnotatorProject, AnnotationChangeHistory, \
-    UserDocumentFormatPreference
-from backend.utils.misc import get_value_from_key_path, insert_value_to_key_path
+    UserDocumentFormatPreference, document_preference_str
+from backend.utils.misc import get_value_from_key_path, insert_value_to_key_path, read_custom_document
 from backend.utils.serialize import ModelSerializer
 
 log = logging.getLogger(__name__)
@@ -34,6 +35,40 @@ log = logging.getLogger(__name__)
 serializer = ModelSerializer()
 User = get_user_model()
 
+#####################################
+### Initilisation                 ###
+#####################################
+@rpc_method
+def initialise(request):
+    """
+    Provide the initial context information to initialise the Teamware app
+
+    context_object:
+        user:
+            isAuthenticated: bool
+            isManager: bool
+            isAdmin: bool
+        configs:
+            docFormatPref: bool
+        global_configs:
+            allowUserDelete: bool
+    """
+    context_object = {
+        "user": is_authenticated(request),
+        "configs": {
+            "docFormatPref": get_user_document_pref_from_request(request)
+        },
+        "global_configs": {
+            "allowUserDelete": settings.ALLOW_USER_DELETE
+        }
+    }
+    return context_object
+
+def get_user_document_pref_from_request(request):
+    if request.user.is_authenticated:
+        return document_preference_str(request.user.doc_format_pref)
+    else:
+        return document_preference_str(UserDocumentFormatPreference.JSON)
 
 #####################################
 ### Login/Logout/Register Methods ###
@@ -75,10 +110,14 @@ def login(request, payload):
 
     user = authenticate(username=payload["username"], password=payload["password"])
     if user is not None:
+
+        if user.is_deleted:
+            raise AuthError("Cannot login with a deleted account")
+
         djlogin(request, user)
         context["username"] = user.username
         context["isAuthenticated"] = user.is_authenticated
-        context["isManager"] = user.is_manager
+        context["isManager"] = user.is_manager or user.is_staff
         context["isAdmin"] = user.is_staff
         context["isActivated"] = user.is_activated
         return context
@@ -98,9 +137,10 @@ def register(request, payload):
     username = payload.get("username")
     password = payload.get("password")
     email = payload.get("email")
+    agreed_privacy_policy = True
 
     if not get_user_model().objects.filter(username=username).exists():
-        user = get_user_model().objects.create_user(username=username, password=password, email=email)
+        user = get_user_model().objects.create_user(username=username, password=password, email=email, agreed_privacy_policy=agreed_privacy_policy)
         _generate_user_activation(user)
         djlogin(request, user)
         context["username"] = payload["username"]
@@ -307,10 +347,7 @@ def get_user_details(request):
     data["user_role"] = user_role
 
     # Convert doc preference to string
-    if user.doc_format_pref == UserDocumentFormatPreference.JSON:
-        data["doc_format_pref"] = "JSON"
-    else:
-        data["doc_format_pref"] = "CSV"
+    data["doc_format_pref"] = document_preference_str(user.doc_format_pref)
 
     return data
 @rpc_method_auth
@@ -367,14 +404,25 @@ def get_user_annotations_in_project(request, project_id, current_page=1, page_si
     else:
         paginated_docs = user_annotated_docs
 
-
     documents_out = []
     for document in paginated_docs:
         annotations_list = [annotation.get_listing() for annotation in document.annotations.filter(user=user)]
         documents_out.append(document.get_listing(annotations_list))
 
-
     return {"items": documents_out, "total_count": total_count}
+
+
+@rpc_method_auth
+def user_delete_personal_information(request):
+    request.user.delete_user_personal_information()
+
+
+@rpc_method_auth
+def user_delete_account(request):
+    if settings.ALLOW_USER_DELETE:
+        request.user.delete()
+    else:
+        raise Exception("Teamware's current configuration does not allow user accounts to be deleted.")
 
 
 ##################################
@@ -415,7 +463,6 @@ def get_project(request, project_id):
         **proj.get_project_stats()
     }
     return out_proj
-
 
 
 @rpc_method_manager
@@ -669,7 +716,7 @@ def get_possible_annotators(request, proj_id):
     active_annotators = User.objects.filter(annotatorproject__status=AnnotatorProject.ACTIVE).values_list('id', flat=True)
     project_annotators = project.annotators.all().values_list('id', flat=True)
     # Do an exclude filter to remove annotator with the those ids
-    valid_annotators = User.objects.exclude(id__in=active_annotators).exclude(id__in=project_annotators)
+    valid_annotators = User.objects.filter(is_deleted=False).exclude(id__in=active_annotators).exclude(id__in=project_annotators)
     output = [serializer.serialize(annotator, {"id", "username", "email"}) for annotator in valid_annotators]
     return output
 
@@ -939,6 +986,47 @@ def admin_update_user_password(request, username, password):
     user = User.objects.get(username=username)
     user.set_password(password)
     user.save()
+
+
+@rpc_method_admin
+def admin_delete_user_personal_information(request, username):
+    user = User.objects.get(username=username)
+    user.delete_user_personal_information()
+
+
+@rpc_method_admin
+def admin_delete_user(request, username):
+    if settings.ALLOW_USER_DELETE:
+        user = User.objects.get(username=username)
+        user.delete()
+    else:
+        raise Exception("Teamware's current configuration does not allow the deleting of users")
+
+
+##################################
+### Privacy Policy/T&C Methods ###
+##################################
+
+@rpc_method
+def get_privacy_policy_details(request):
+    details = settings.PRIVACY_POLICY
+
+    custom_docs = {
+        'CUSTOM_PP_DOCUMENT': read_custom_document(settings.CUSTOM_PP_DOCUMENT_PATH) if os.path.isfile(
+            settings.CUSTOM_PP_DOCUMENT_PATH) else None,
+        'CUSTOM_TC_DOCUMENT': read_custom_document(settings.CUSTOM_TC_DOCUMENT_PATH) if os.path.isfile(
+            settings.CUSTOM_TC_DOCUMENT_PATH) else None
+    }
+
+    details.update(custom_docs)
+
+    url = {
+        'URL': request.headers['Host']
+    }
+
+    details.update(url)
+
+    return details
 
 
 ###############################