+++ deploy.sh
@@ -6,14 +6,9 @@ source .env
 
 DEPLOY_ENV=$1
 
-DJANGO_SETTINGS_MODULE=teamware.settings.deployment
-export DJANGO_SETTINGS_MODULE
-
 case $DEPLOY_ENV in
 
   production|prod)
-    DJANGO_ALLOWED_HOSTS=$TEAMWARE_HOST_URL_PRODUCTION
-    export DJANGO_ALLOWED_HOSTS
     echo "Deploying with DJANGO_ALLOWED_HOSTS: $DJANGO_ALLOWED_HOSTS"
     ;;