+++ docs/docs.config.js
@@ -1,9 +1,9 @@
 module.exports = {
-    documentationDir: "docs/docs",
-    documentationVersionsDir: "docs/versioned",
-    buildTargetDir:"docs/site/gate-teamware/",
+    documentationDir: "docs",
+    documentationVersionsDir: "versioned",
+    buildTargetDir:"site/gate-teamware/",
     base:"/gate-teamware/",
-    frontendSourceDir:"frontend/src",
+    frontendSourceDir:"../frontend/src",
     latestVersionName: "development",
     defaultVersion:"development"
 }