+++ frontend/src/components/navbar.vue
@@ -27,7 +27,6 @@
               <b-icon-box-arrow-up-right
                   style="position: relative; font-size: 0.8em; padding-bottom: 0.2em;"></b-icon-box-arrow-up-right>
             </b-nav-item>
-            <b-nav-item to="/about">About</b-nav-item>
           </b-navbar-nav>